Split up into separate pakages (#50)
This is a re-structuring. The functionality remains the same.

a base stackfmt package. This package is just for stack traces and does not do anything with errors.
A separate slogerr package for those looking to add structured errors.
An errwrap package for wrapping and unwrapping errors
Maintain the existing errors package. This is designed as a replacement for the existing errors package that introduces wrapping with stack traces
The errors package will re-export some items from other packages.
However, it is exporting much less overall.